http://mobileappchallenge.ucmerced.edu/projects/true-posture
According to a study by the Centers for Disease Control and Prevention
from 2005, back pain is one of the three most commons causes of
disability, affecting a total of 7.6 million of people only in the U.S.
Back pain is caused by bad posture which is also known as a
musculoskeletal disorder, injuries or pain in the body's joints,
ligaments, muscles, nerves, tendons, and structures that support limbs,
neck and back. For the average American, who remains in a sitting or
standing position at an office for 8 hours, 5 days every week bad
posture can become easily a routine leading to serious health problems.
True Posture will prevent back pain by advising the user on how to keep
a correct posture.

By preventing back pain, the individual does not only prevent future
health problems but it will also help the health system saving millions
of dollars in treatments and medications, money that can be used for
people with disorders that cannot be prevented as easily such as back
pain. Overall, True Posture will help people correct their posture
before it becomes a potential health problem that can lead to expensive
treatments, medications, and devices.

Our app is going to connect to a wearable device that takes data on your
current posture. Once the user starts the applications and connects it
to the device via Bluetooth, the app will collect data and save that
data for the user to view later. With the saved data, the app will
display a graph of the data to show how your posture looks over a set
period of time. The most recent data can be compared to old data of your
posture and show how much you have improved your posture since you
started using the device.

In addition to keeping track of your posture, the app will be able to
notify the user when they are practicing bad posture. When the app is in
its “alert” mode, the app will run as a background process and send the
user a notification or alarm that they are practicing bad posture. Also,
the user can set how long bad posture is being practiced before they get
notified. For example, if the user is searching for an immediate
response, the app can notify them when their posture is bad for more
than five seconds. Thus, this apps main function is to monitor the daily
person’s posture and improve it.